Humble beginnings for a new design
Okay, so:
- New logo
- New framework: From now on, I want to use Fitgrid, which has all the
layout stuff we need and none of the stuff we don’t. Anything extra
(such as Javascript) we can take as needed from Bootstrap. Doing it
piecemeal makes way more sense than just downloading the whole
Bootstrap, which is really long and a royal pain in the ass to
modify/hack.
- Panes are as follows:
1) Banner
2) Carousel, Description, Press
3) Near You, Side-By-Side
4) Have a Conversation
5) Flowchart (escrow)
6) Cloud Storage
7) Big Photo (the fluffy feel-good moment)
8) Contact
9) Footer
- Lots of subtle, classy CSS/JS animations on the way.

Please note that this design is very rough around the edges still, and
it’s very very hungry for critical feedback and suggestions. Please
feed it.