New fonts, prettier backgrounds, and cleaner CSS
We can’t use DIN Next unless someone wants to pay a ton. “Lato” is a
Google Font that I found to be pretty similar and really kickass.

Added pink pane and little hearts for texture.

The h1/h2/h3 hierarchy was getting really messy. This is way more
straightforward